hook into docstring
