+++ sphinx_gallery/tests/test_gen_gallery.py
@@ -14,8 +14,7 @@ import pytest
 from sphinx.errors import ConfigError, ExtensionError, SphinxWarning
 from sphinx_gallery.gen_gallery import (
     check_duplicate_filenames, check_spaces_in_filenames,
-    collect_gallery_files, write_computation_times, _complete_gallery_conf,
-    write_api_entry_usage)
+    collect_gallery_files, write_computation_times, _complete_gallery_conf)
 
 
 def test_bad_config():
@@ -460,10 +459,6 @@ def test_backreferences_dir_pathlib_config(sphinx_app_wrapper):
 
 def test_write_computation_times_noop():
     write_computation_times(None, None, [[[0]]])
-
-
-def test_write_api_usage_noop():
-    write_api_entry_usage({'backreferences_dir': None}, None)
 
 
 @pytest.mark.conf_file(content="""
